+++ src/rules/enforce-relation-types.test.ts
@@ -14,23 +14,51 @@ const ruleTester = new RuleTester({
 ruleTester.run('enforce-relation-types', enforceRelationTypes, {
     valid: [
         {
-            name: 'should allow valid one-to-one relations',
+            name: 'should allow valid nullable one-to-one relations',
             code: `class Entity {
                 @OneToOne(() => Other)
                 @JoinColumn()
                 other: Other | null;
-
+            }`,
+        },
+        {
+            name: 'should allow valid nullable one-to-one relations with relation defined',
+            code: `class Entity {
                 @OneToOne(() => Other, (other) => other.entity)
                 @JoinColumn()
                 other: Other | null;
-                
+            }`,
+        },
+        {
+            name: 'should allow valid non-nullable one-to-one relations',
+            code: `class Entity {
                 @OneToOne(() => Other, { nullable: false })
                 @JoinColumn()
                 other: Other;
-
+            }`,
+        },
+        {
+            name: 'should allow valid non-nullable one-to-one relations with relation defined',
+            code: `class Entity {
                 @OneToOne(() => Other, (other) => other.entity, { nullable: false })
                 @JoinColumn()
                 other: Other;
+            }`,
+        },
+        {
+            name: 'should allow valid lazy one-to-one relations',
+            code: `class Entity {
+                @OneToOne(() => Other)
+                @JoinColumn()
+                other: Promise<Other | null>;
+            }`,
+        },
+        {
+            name: 'should allow valid lazy non-nullable one-to-one relations',
+            code: `class Entity {
+                @OneToOne(() => Other, { nullable: false })
+                @JoinColumn()
+                other: Promise<Other>;
             }`,
         },
         {
@@ -41,19 +69,52 @@ ruleTester.run('enforce-relation-types', enforceRelationTypes, {
             }`,
         },
         {
-            name: 'should allow valid many-to-one relations',
+            name: 'should allow lazy one-to-many relations',
+            code: `class Entity {
+                @OneToMany(() => Other, (other) => other.entity)
+                others: Promise<Other[]>;
+            }`,
+        },
+        {
+            name: 'should allow valid nullable many-to-one relations',
             code: `class Entity {
                 @ManyToOne(() => Other)
                 other: Other | null;
-
+            }`,
+        },
+        {
+            name: 'should allow valid non-nullable many-to-one relations',
+            code: `class Entity {
                 @ManyToOne(() => Other, { nullable: false })
                 other: Other;
-                
+            }`,
+        },
+        {
+            name: 'should allow valid nullable many-to-one relations with relation defined',
+            code: `class Entity {
                 @ManyToOne(() => Other, (other) => other.entity)
                 other: Other | null;
-
+            }`,
+        },
+        {
+            name: 'should allow valid non-nullable many-to-one relations with relation defined',
+            code: `class Entity {
                 @ManyToOne(() => Other, (other) => other.entity, { nullable: false })
                 other: Other;
+            }`,
+        },
+        {
+            name: 'should allow valid lazy many-to-one relations',
+            code: `class Entity {
+                @ManyToOne(() => Other)
+                other: Promise<Other | null>;
+            }`,
+        },
+        {
+            name: 'should allow valid lazy non-nullable many-to-one relations',
+            code: `class Entity {
+                @ManyToOne(() => Other, { nullable: false })
+                other: Promise<Other>;
             }`,
         },
         {
@@ -64,8 +125,29 @@ ruleTester.run('enforce-relation-types', enforceRelationTypes, {
                 others: Other[];
             }`,
         },
+        {
+            name: 'should allow lazy many-to-many relations',
+            code: `class Entity {
+                @ManyToMany(() => Other)
+                @JoinTable()
+                others: Promise<Other[]>;
+            }`,
+        },
     ],
     invalid: [
+        {
+            name: 'should fail on nullable one-to-one relations',
+            code: `class Entity {
+                @OneToOne()
+                @JoinColumn()
+                other: Other;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_missing',
+                },
+            ],
+        },
         {
             name: 'should fail on nullable one-to-one relations',
             code: `class Entity {
@@ -137,6 +219,29 @@ ruleTester.run('enforce-relation-types', enforceRelationTypes, {
                 @OneToOne(() => Other)
                 @JoinColumn()
                 other: Other | null;
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
+            name: 'should fail on mismatched one-to-one promise relations',
+            code: `class Entity {
+                @OneToOne(() => Other)
+                @JoinColumn()
+                other: Promise<Another | null>;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_mismatch',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @OneToOne(() => Other)
+                @JoinColumn()
+                other: Promise<Other | null>;
             }`,
                         },
                     ],
@@ -255,6 +360,27 @@ ruleTester.run('enforce-relation-types', enforceRelationTypes, {
             ],
         },
         {
+            name: 'should fail on mismatched lazy one-to-many relations',
+            code: `class Entity {
+                    @OneToMany(() => Other, (other) => other.entity)
+                    others: Promise<Another[]>;
+                }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_mismatch',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                    @OneToMany(() => Other, (other) => other.entity)
+                    others: Promise<Other[]>;
+                }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
             name: 'should fail on primitive one-to-many relations',
             code: `class Entity {
                 @OneToMany(() => Other, (other) => other.entity)
@@ -346,6 +472,27 @@ ruleTester.run('enforce-relation-types', enforceRelationTypes, {
             ],
         },
         {
+            name: 'should fail on mismatched many-to-one relations',
+            code: `class Entity {
+                @ManyToOne(() => Other)
+                other: Promise<Another | null>;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_mismatch',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @ManyToOne(() => Other)
+                other: Promise<Other | null>;
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
             name: 'should fail on primitive many-to-one relations',
             code: `class Entity {
                 @ManyToOne(() => Other)
@@ -427,6 +574,29 @@ ruleTester.run('enforce-relation-types', enforceRelationTypes, {
                 @ManyToMany(() => Other)
                 @JoinTable()
                 others: Other[];
+            }`,
+                        },
+                    ],
+                },
+            ],
+        },
+        {
+            name: 'should fail on mismatched lazy many-to-many relations',
+            code: `class Entity {
+                @ManyToMany(() => Other)
+                @JoinTable()
+                others: Promise<Another[]>;
+            }`,
+            errors: [
+                {
+                    messageId: 'typescript_typeorm_relation_mismatch',
+                    suggestions: [
+                        {
+                            messageId: 'typescript_typeorm_relation_suggestion',
+                            output: `class Entity {
+                @ManyToMany(() => Other)
+                @JoinTable()
+                others: Promise<Other[]>;
             }`,
                         },
                     ],