+++ src/utils/relationType.ts
@@ -5,6 +5,7 @@ interface RelationType {
     name: string;
     nullable: boolean;
     isArray: boolean;
+    isLazy: boolean;
 }
 
 export type Relation = 'OneToOne' | 'OneToMany' | 'ManyToOne' | 'ManyToMany';
@@ -12,9 +13,18 @@ export type Relation = 'OneToOne' | 'OneToMany' | 'ManyToOne' | 'ManyToMany';
 export function convertTypeToRelationType(arg: TSESTree.TypeNode): RelationType {
     switch (arg.type) {
         case AST_NODE_TYPES.TSTypeReference: {
+            const name = arg.typeName.type === AST_NODE_TYPES.Identifier ? arg.typeName.name : '';
+            const param = arg.typeArguments?.params?.[0];
+            if (name === 'Promise' && param) {
+                return {
+                    ...convertTypeToRelationType(param),
+                    isLazy: true,
+                };
+            }
             return {
-                name: arg.typeName.type === AST_NODE_TYPES.Identifier ? arg.typeName.name : '',
+                name,
                 isArray: false,
+                isLazy: false,
                 nullable: false,
             };
         }
@@ -23,7 +33,7 @@ export function convertTypeToRelationType(arg: TSESTree.TypeNode): RelationType 
             return { ...item, isArray: true };
         }
         case AST_NODE_TYPES.TSNullKeyword: {
-            return { name: '', isArray: false, nullable: true };
+            return { name: '', isArray: false, isLazy: false, nullable: true };
         }
         case AST_NODE_TYPES.TSUnionType: {
             return arg.types.reduce(
@@ -32,29 +42,28 @@ export function convertTypeToRelationType(arg: TSESTree.TypeNode): RelationType 
                     return {
                         name: acc.name || current.name,
                         isArray: acc.isArray || current.isArray,
+                        isLazy: acc.isLazy || current.isLazy,
                         nullable: acc.nullable || current.nullable,
                     };
                 },
                 {
                     name: '',
                     isArray: false,
+                    isLazy: false,
                     nullable: false,
                 } as RelationType,
             );
         }
         default: {
-            return { name: '', isArray: false, nullable: false };
+            return { name: '', isArray: false, isLazy: false, nullable: false };
         }
     }
 }
 
 export function convertArgumentToRelationType(
     relation: Relation,
-    args: TSESTree.CallExpressionArgument[] | undefined,
+    args: TSESTree.CallExpressionArgument[],
 ): RelationType | undefined {
-    if (!args) {
-        return undefined;
-    }
     const [otherEntityFn, ...restArguments] = args;
     const otherEntity = findReturnedValue(otherEntityFn);
     if (!otherEntity) {
@@ -63,24 +72,22 @@ export function convertArgumentToRelationType(
 
     // OneToMany, ManyToMany
     if (relation === 'OneToMany' || relation === 'ManyToMany') {
-        return { name: otherEntity, isArray: true, nullable: false };
+        return { name: otherEntity, isArray: true, isLazy: false, nullable: false };
     }
     // OneToOne, ManyToOne
     const options = findObjectArgument(restArguments);
     const parsedOptions = parseObjectLiteral(options) as { nullable?: boolean } | undefined;
 
-    return { name: otherEntity, isArray: false, nullable: parsedOptions?.nullable ?? true };
+    return {
+        name: otherEntity,
+        isArray: false,
+        isLazy: false,
+        nullable: parsedOptions?.nullable ?? true,
+    };
 }
 
-export function isTypesEqual(
-    toType: RelationType | undefined,
-    tsType: RelationType | undefined,
-): boolean {
-    // If either is undefined, that means we are not sure of the types... ignore
-    if (!toType || !tsType) {
-        return true;
-    }
-    // Otherwise just check field equality
+export function isTypesEqual(toType: RelationType, tsType: RelationType): boolean {
+    // Just check field equality
     return (
         toType.name === tsType.name &&
         toType.nullable === tsType.nullable &&
@@ -89,42 +96,27 @@ export function isTypesEqual(
 }
 
 // Relations are nullable by default which can be confusing, help with a custom message
-export function isTypeMissingNullable(
-    toType: RelationType | undefined,
-    tsType: RelationType | undefined,
-): boolean {
-    // If either is undefined, that means we are not sure of the types... ignore
-    if (!toType || !tsType) {
-        return false;
-    }
-    // Otherwise check if the relation should be nullable but not
+export function isTypeMissingNullable(toType: RelationType, tsType: RelationType): boolean {
+    // Check if the relation should be nullable but not
     return toType.name === tsType.name && toType.nullable && !tsType.nullable;
 }
 
 // Relations are nullable by default which can be confusing, help with a custom message
-export function isTypeMissingArray(
-    toType: RelationType | undefined,
-    tsType: RelationType | undefined,
-): boolean {
-    // If either is undefined, that means we are not sure of the types... ignore
-    if (!toType || !tsType) {
-        return false;
-    }
-    // Otherwise check if the relation should be an array but not
+export function isTypeMissingArray(toType: RelationType, tsType: RelationType): boolean {
+    // Check if the relation should be an array but not
     return toType.name === tsType.name && toType.isArray && !tsType.isArray;
 }
 
-export function typeToString(relation: RelationType | undefined): string | undefined {
-    // If unknown, we don't suggest change
-    if (!relation) {
-        return undefined;
-    }
+export function typeToString(relation: RelationType, { isLazy }: RelationType): string | undefined {
     let result = relation.name;
     if (relation.isArray) {
         result += '[]';
     }
     if (relation.nullable) {
         result += ' | null';
+    }
+    if (isLazy) {
+        result = `Promise<${result}>`;
     }
     return result;
 }